fix(a2-audit): Paso 4 - migrar 5 casos CLP en Recipes (Inventory preservado)
FASE A2 - PASO 4: Análisis conservador de Inventory + Recipes.

CAMBIOS EN CÓDIGO (5 cambios en 4 archivos):

📦 RECIPES (5 cambios):
   - ProductRecipeResource: product_base_price como int
   - ProductRecipe: product->base_price como int (2 ocurrencias)
   - RecipeService: product_base_price como int
   - IngredientController: total_purchase_cost como int

📦 INVENTORY (0 cambios):
   - cost_price: MANTENIDO float (costo unitario con decimales legítimos)
   - quantity, min_stock, balance_after: MANTENIDOS float (cantidades físicas)

CASOS PRESERVADOS COMO FLOAT LEGÍTIMO (54 casos):

1. CANTIDADES FÍSICAS (~30 casos):
   - quantity, current_stock_base, minimum_stock_base
   - purchase_quantity, total_base_quantity_added
   - effective_discount_base_quantity
   → Razón: kg, litros, gramos requieren decimales (ej: 0.250 kg)

2. COSTOS UNITARIOS CON DECIMALES (~10 casos):
   - cost_per_base_unit, cost_price, calculated_item_cost
   → Razón: costo por gramo puede ser $2.5/g (decimal legítimo)

3. PORCENTAJES (~6 casos):
   - waste_percentage, food_cost_percentage
   → Razón: porcentajes requieren decimales (ej: 15.5%)

4. RATIOS DE CONVERSIÓN (~2 casos):
   - conversion_factor_to_base
   → Razón: factor de conversión (ej: 1000 g/kg)

5. VALORES DERIVADOS (~6 casos):
   - total_recipe_cost, gross_margin, total_stock_value
   → Razón: resultan de multiplicar cantidad × costo unitario,
     heredan decimales. Migrar requeriría redondear y perder
     precisión en análisis de food cost.

DECISIÓN ARQUITECTÓNICA:
ADR-018 aplica a la CADENA DE PAGOS (Orders, Bills, Payments, Cashier).
Los costos de inventario y recetas son para ANÁLISIS INTERNO (food cost)
y pueden mantener float legítimamente para preservar precisión.

PROGRESO ACUMULADO FASE A2:
✅ Paso 1: 13 casos
✅ Paso 2: 18 casos
✅ Paso 3: 39 casos
✅ Paso 4: 5 casos
📊 Total: 75/143 casos migrados (~52%)
📊 Float legítimo documentado: ~68 casos

RESULTADO:
✅ 1021 tests passing, 0 failed
✅ Consistencia con ADR-018 en cadena de pagos
✅ Preservada precisión en análisis de food cost

FASE A2 COMPLETADA: Todos los montos CLP de la cadena de pagos
son integer. Los floats residuales son legítimos y documentados.

+++ app/Modules/Recipes/Domain/Entities/ProductRecipe.php
@@ -70,7 +70,7 @@ class ProductRecipe extends Model
      */
     public function calculateFoodCostPercentage(): float
     {
-        $productPrice = (float) ($this->product?->base_price ?? 0);
+        $productPrice = (int) ($this->product?->base_price ?? 0);  // ADR-018
         if ($productPrice <= 0) {
             return 0.0;
         }
@@ -85,7 +85,7 @@ class ProductRecipe extends Model
      */
     public function calculateGrossMargin(): float
     {
-        $productPrice = (float) ($this->product?->base_price ?? 0);
+        $productPrice = (int) ($this->product?->base_price ?? 0);  // ADR-018
         return max(0, $productPrice - (float) $this->total_recipe_cost);
     }
 }
